refactor: extract custom hooks for comment-list and comment-write
