+++ 03-maincamp-triptalk/03-CNA/src/components/boards-detail/comment-list/hook.ts
@@ -1,0 +1,20 @@
+"use client";
+
+import { useQuery } from "@apollo/client";
+import { useParams } from "next/navigation";
+import { FETCH_BOARD_COMMENTS } from "./queries";
+
+export default function useCommentList() {
+  const { boardId } = useParams<{ boardId: string }>();
+
+  const { data, loading, error, refetch } = useQuery(FETCH_BOARD_COMMENTS, {
+    variables: { boardId },
+    skip: !boardId,
+  });
+
+  const comments = data?.fetchBoardComments ?? [];
+
+  return { comments, loading, error, refetch };
+}
+
+
